[linkis-storage] added license
